Clean up seeding code: - Move to own service - Remove controller endpoint so it's not publicly accessible
