+++ BrowserVersions.API/Services/IBrowserVersionSeedingService.cs
@@ -1,0 +1,7 @@
+namespace BrowserVersions.API.Services {
+  using System.Threading.Tasks;
+
+  public interface IBrowserVersionSeedingService {
+    Task SeedBrowserData();
+  }
+}
