+++ BrowserVersions.API/Services/IBrowserVersionService.cs
@@ -7,7 +7,5 @@ namespace BrowserVersions.API.Services {
 
   public interface IBrowserVersionService {
     Task<Dictionary<TargetBrowser, Dictionary<Platform, VersionChannels>>> GetBrowserVersion(List<TargetBrowser> targetBrowsers, List<Platform> targetPlatforms, DateTime? minSupportDate, DateTime? maxSupportDate);
-
-    Task AddHistoricalData();
   }
 }
